Create README.md
Building a face recognition program with the help of python and its libraries such as NumPy and OpenCV. Once the program is trained on the data(images of the face of a person), it will be able to recognize that person. The data is processed using K-Nearest Neighbour Clustering Algorithm.